Resampler: Use float instead of double for coefficient calculation. Faster and good enough precision.

+++ plugins/resample/resample.c
@@ -54,7 +54,7 @@ typedef struct {
 	guint dest_offset_other;	/* Where in the unchanged direction should we begin writing? */
 	guint dest_end_other;		/* Where in the unchanged direction should we stop writing? */
 	guint (*resample_support)();
-	gdouble (*resample_func)(gdouble);
+	gfloat (*resample_func)(gfloat);
 	GThread *threadid;
 	gboolean use_compatible;	/* Use compatible resampler if pixelsize != 4 */
 	gboolean use_fast;		/* Use nearest neighbour resampler, also compatible*/
@@ -446,28 +446,28 @@ lanczos_taps()
 	return 3;
 }
 
-static gdouble
-sinc(gdouble value)
+static gfloat
+sinc(gfloat value)
 {
-	if (value != 0.0)
+	if (value != 0.0f)
 	{
 		value *= M_PI;
-		return sin(value) / value;
+		return sinf(value) / value;
 	}
 	else
-		return 1.0;
+		return 1.0f;
 }
 
-static gdouble
-lanczos_weight(gdouble value)
+static gfloat
+lanczos_weight(gfloat value)
 {
-	value = fabs(value);
+	value = fabsf(value);
 	if (value < lanczos_taps())
 	{
 		return (sinc(value) * sinc(value / lanczos_taps()));
 	}
 	else
-		return 0.0;
+		return 0.0f;
 }
 
 const static gint FPScale = 16384; /* fixed point scaler */
@@ -481,9 +481,9 @@ ResizeH(ResampleInfo *info)
 	const guint old_size = info->old_size;
 	const guint new_size = info->new_size;
 
-	gdouble pos_step = ((gdouble) old_size) / ((gdouble)new_size);
-	gdouble filter_step = MIN(1.0 / pos_step, 1.0);
-	gdouble filter_support = (gdouble) lanczos_taps() / filter_step;
+	gfloat pos_step = ((gfloat) old_size) / ((gfloat)new_size);
+	gfloat filter_step = MIN(1.0 / pos_step, 1.0);
+	gfloat filter_support = (gfloat) lanczos_taps() / filter_step;
 	gint fir_filter_size = (gint) (ceil(filter_support*2));
 
 	if (old_size <= fir_filter_size)
@@ -492,7 +492,7 @@ ResizeH(ResampleInfo *info)
 	gint *weights = g_new(gint, new_size * fir_filter_size);
 	gint *offsets = g_new(gint, new_size);
 
-	gdouble pos = 0.0;
+	gfloat pos = 0.0f;
 	gint i,j,k;
 
 	for (i=0; i<new_size; ++i)
@@ -510,10 +510,10 @@ ResizeH(ResampleInfo *info)
 		offsets[i] = start_pos * 4;
 
 		/* the following code ensures that the coefficients add to exactly FPScale */
-		gdouble total = 0.0;
+		gfloat total = 0.0;
 
 		/* Ensure that we have a valid position */
-		gdouble ok_pos = MAX(0.0,MIN(old_size-1,pos));
+		gfloat ok_pos = MAX(0.0,MIN(old_size-1,pos));
 
 		for (j=0; j<fir_filter_size; ++j)
 		{
@@ -523,11 +523,11 @@ ResizeH(ResampleInfo *info)
 
 		g_assert(total > 0.0f);
 
-		gdouble total2 = 0.0;
+		gfloat total2 = 0.0;
 
 		for (k=0; k<fir_filter_size; ++k)
 		{
-			gdouble total3 = total2 + lanczos_weight((start_pos+k - ok_pos) * filter_step) / total;
+			gfloat total3 = total2 + lanczos_weight((start_pos+k - ok_pos) * filter_step) / total;
 			weights[i*fir_filter_size+k] = (gint) (total3*FPScale+0.5) - (gint) (total2*FPScale+0.5);
 			total2 = total3;
 		}
@@ -580,9 +580,9 @@ ResizeV(ResampleInfo *info)
 	const guint start_x = info->dest_offset_other;
 	const guint end_x = info->dest_end_other;
 
-	gdouble pos_step = ((gdouble) old_size) / ((gdouble)new_size);
-	gdouble filter_step = MIN(1.0 / pos_step, 1.0);
-	gdouble filter_support = (gdouble) lanczos_taps() / filter_step;
+	gfloat pos_step = ((gfloat) old_size) / ((gfloat)new_size);
+	gfloat filter_step = MIN(1.0 / pos_step, 1.0);
+	gfloat filter_support = (gfloat) lanczos_taps() / filter_step;
 	gint fir_filter_size = (gint) (ceil(filter_support*2));
 
 	if (old_size <= fir_filter_size)
@@ -591,7 +591,7 @@ ResizeV(ResampleInfo *info)
 	gint *weights = g_new(gint, new_size * fir_filter_size);
 	gint *offsets = g_new(gint, new_size);
 
-	gdouble pos = 0.0;
+	gfloat pos = 0.0;
 
 	gint i,j,k;
 
@@ -609,10 +609,10 @@ ResizeV(ResampleInfo *info)
 		offsets[i] = start_pos;
 
 		/* The following code ensures that the coefficients add to exactly FPScale */
-		gdouble total = 0.0;
+		gfloat total = 0.0;
 
 		/* Ensure that we have a valid position */
-		gdouble ok_pos = MAX(0.0,MIN(old_size-1,pos));
+		gfloat ok_pos = MAX(0.0,MIN(old_size-1,pos));
 
 		for (j=0; j<fir_filter_size; ++j)
 		{
@@ -622,11 +622,11 @@ ResizeV(ResampleInfo *info)
 
 		g_assert(total > 0.0f);
 
-		gdouble total2 = 0.0;
+		gfloat total2 = 0.0;
 
 		for (k=0; k<fir_filter_size; ++k)
 		{
-			gdouble total3 = total2 + lanczos_weight((start_pos+k - ok_pos) * filter_step) / total;
+			gfloat total3 = total2 + lanczos_weight((start_pos+k - ok_pos) * filter_step) / total;
 			weights[i*fir_filter_size+k] = (gint) (total3*FPScale+0.5) - (gint) (total2*FPScale+0.5);
 			total2 = total3;
 		}
@@ -678,9 +678,9 @@ ResizeH_compatible(ResampleInfo *info)
 	gint pixelsize = input->pixelsize;
 	gint ch = input->channels;
 
-	gdouble pos_step = ((gdouble) old_size) / ((gdouble)new_size);
-	gdouble filter_step = MIN(1.0 / pos_step, 1.0);
-	gdouble filter_support = (gdouble) lanczos_taps() / filter_step;
+	gfloat pos_step = ((gfloat) old_size) / ((gfloat)new_size);
+	gfloat filter_step = MIN(1.0 / pos_step, 1.0);
+	gfloat filter_support = (gfloat) lanczos_taps() / filter_step;
 	gint fir_filter_size = (gint) (ceil(filter_support*2));
 
 	if (old_size <= fir_filter_size)
@@ -689,7 +689,7 @@ ResizeH_compatible(ResampleInfo *info)
 	gint *weights = g_new(gint, new_size * fir_filter_size);
 	gint *offsets = g_new(gint, new_size);
 
-	gdouble pos = 0.0;
+	gfloat pos = 0.0;
 	gint i,j,k;
 
 	for (i=0; i<new_size; ++i)
@@ -707,10 +707,10 @@ ResizeH_compatible(ResampleInfo *info)
 		offsets[i] = start_pos * pixelsize;
 
 		/* the following code ensures that the coefficients add to exactly FPScale */
-		gdouble total = 0.0;
+		gfloat total = 0.0;
 
 		/* Ensure that we have a valid position */
-		gdouble ok_pos = MAX(0.0,MIN(old_size-1,pos));
+		gfloat ok_pos = MAX(0.0,MIN(old_size-1,pos));
 
 		for (j=0; j<fir_filter_size; ++j)
 		{
@@ -720,11 +720,11 @@ ResizeH_compatible(ResampleInfo *info)
 
 		g_assert(total > 0.0f);
 
-		gdouble total2 = 0.0;
+		gfloat total2 = 0.0;
 
 		for (k=0; k<fir_filter_size; ++k)
 		{
-			gdouble total3 = total2 + lanczos_weight((start_pos+k - ok_pos) * filter_step) / total;
+			gfloat total3 = total2 + lanczos_weight((start_pos+k - ok_pos) * filter_step) / total;
 			weights[i*fir_filter_size+k] = (gint) (total3*FPScale+0.5) - (gint) (total2*FPScale+0.5);
 			total2 = total3;
 		}
@@ -775,9 +775,9 @@ ResizeV_compatible(ResampleInfo *info)
 	gint pixelsize = input->pixelsize;
 	gint ch = input->channels;
 
-	gdouble pos_step = ((gdouble) old_size) / ((gdouble)new_size);
-	gdouble filter_step = MIN(1.0 / pos_step, 1.0);
-	gdouble filter_support = (gdouble) lanczos_taps() / filter_step;
+	gfloat pos_step = ((gfloat) old_size) / ((gfloat)new_size);
+	gfloat filter_step = MIN(1.0 / pos_step, 1.0);
+	gfloat filter_support = (gfloat) lanczos_taps() / filter_step;
 	gint fir_filter_size = (gint) (ceil(filter_support*2));
 
 	if (old_size <= fir_filter_size)
@@ -786,7 +786,7 @@ ResizeV_compatible(ResampleInfo *info)
 	gint *weights = g_new(gint, new_size * fir_filter_size);
 	gint *offsets = g_new(gint, new_size);
 
-	gdouble pos = 0.0;
+	gfloat pos = 0.0;
 
 	gint i,j,k;
 
@@ -804,10 +804,10 @@ ResizeV_compatible(ResampleInfo *info)
 		offsets[i] = start_pos;
 
 		/* The following code ensures that the coefficients add to exactly FPScale */
-		gdouble total = 0.0;
+		gfloat total = 0.0;
 
 		/* Ensure that we have a valid position */
-		gdouble ok_pos = MAX(0.0,MIN(old_size-1,pos));
+		gfloat ok_pos = MAX(0.0,MIN(old_size-1,pos));
 
 		for (j=0; j<fir_filter_size; ++j)
 		{
@@ -817,11 +817,11 @@ ResizeV_compatible(ResampleInfo *info)
 
 		g_assert(total > 0.0f);
 
-		gdouble total2 = 0.0;
+		gfloat total2 = 0.0;
 
 		for (k=0; k<fir_filter_size; ++k)
 		{
-			gdouble total3 = total2 + lanczos_weight((start_pos+k - ok_pos) * filter_step) / total;
+			gfloat total3 = total2 + lanczos_weight((start_pos+k - ok_pos) * filter_step) / total;
 			weights[i*fir_filter_size+k] = (gint) (total3*FPScale+0.5) - (gint) (total2*FPScale+0.5);
 			total2 = total3;
 		}
@@ -868,7 +868,7 @@ ResizeV_fast(ResampleInfo *info)
 	gint pixelsize = input->pixelsize;
 	gint ch = input->channels;
 
-	gdouble pos_step = ((gdouble) old_size) / ((gdouble)new_size);
+	gfloat pos_step = ((gfloat) old_size) / ((gfloat)new_size);
 
 	gint pos = 0;
 	gint delta = (gint)(pos_step * 65536.0);
@@ -903,7 +903,7 @@ ResizeH_fast(ResampleInfo *info)
 	gint pixelsize = input->pixelsize;
 	gint ch = input->channels;
 
-	gdouble pos_step = ((gdouble) old_size) / ((gdouble)new_size);
+	gfloat pos_step = ((gfloat) old_size) / ((gfloat)new_size);
 
 	gint pos;
 	gint delta = (gint)(pos_step * 65536.0);
